refactor(sysctl): add hpm6e support, refine sub mod structure
